+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - Fixed issue with monorepos using Rush (#667)
 - Fixed issue with blocked port in case of full debug reload (#671)
 - Fixed `@Inject` with multiple instances in `piral-ng` (#673)
+- Fixed standalone pilets depending on `@angular/compiler`
 - Removed legacy option in `piral-cli-webpack5` to support IE8
 - Removed pilet-related options in debug settings when running `piral debug` (#670)
 - Improved internal navigation Blazor pilets using `piral-blazor`
 - Updated dependencies
+- Updated `kras` to `v0.16` to fix some issues and remove `request`
 - Updated `piral publish` command to work exclusively for emulator websites
 - Added special entry point to emulator website when accessed online (#654)
 - Added feed selection view for remote emulator website (#654)